weather control, time contrl, and refactored parametre settings

+++ src/main/java/com/hadroncfy/sreplay/config/Config.java
@@ -25,8 +25,8 @@ public class Config {
     public String serverHostName = "localhost";
     public int serverPort = 12346;
     public long downloadTimeout = 60000;
-    public long sizeLimit = -1;
-    public long timeLimit = -1;
+    public int sizeLimit = -1;
+    public int timeLimit = -1;
     public boolean autoReconnect = true;
     public Pattern playerNamePattern = Pattern.compile("^cam_.*$");
